treat it as root node when parentId is empty string
One node should also be treated as root node when its parentId is an empty string. because FetchJsonTreeCommandHandler.java use fastjson to serialize tree node data, if WriteNullStringAsEmpty option in fastjson is set then null string value will be output as empty string, which will convert parentId of root node  to empty string, so this real root node can not be treated as root node.