Update one method name.
The old name is a little bit unclear. So in fact in this method, it is tracking from the entry to exit in order to catch the error messages. But the old name makes people may think that this method is tracking the exception happening process. Just want to make it clear