OSPP2025 - Research and Implementation of Adaptive Circuit Breaker Function The first implementation version of the adaptive circuit breaker function has been corrected.
Revised the log recording levels of some classes.

Added the filtering and extraction of abnormal values when dealing with overload scenarios, and moved the judgment method to the utility class.

Previously, the adaptive circuit-breaking slot was not included in the responsibility chain by default; however, when users manually configured it to be added to the chain, the adaptive circuit-breaking feature would be enabled by default. This design has now been changed: the adaptive circuit-breaking slot is included in the responsibility chain by default, but the feature itself remains disabled by default. Users can enable or disable it dynamically at runtime (i.e., achieve a "hot switch") by registering a property listener (e.g., via Nacos).

Based on the aforementioned changes, some of the test contents of the test classes have been corrected.